requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
supabase==2.3.4
httpx>=0.28.0
